+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 - remove Python 2 support
 - add Model.getParams that returns a dict mapping all parameter names to their values
 - add Model.setParams to set multiple parameters at once using a dict
+- Add Model.from_ptr and Model.to_ptr to interface with SCIP* managed outside of PySCIPOpt
 - Node.getParent() returns None if the node has no parent
+- add NULL pointer checks to all Python wrapper classes
+- add Event.getRow() and Row.name
 - define Python object identity based on underlying SCIP object pointers, so that e.g. rows and columns can be added to sets, and testing for equality is consistent over time.
 
 ## 2.2.3 - 2019-12-10